refactor: move the vault-config parser to shared/lib/money
Upstreaming the parser to core was rejected (clients may legitimately
diverge on what they accept from the flag), but it lived in ui/selectors
where the background cannot reach it — import-x/no-restricted-paths bars
imports between the two layers in either direction. shared/lib/money is the
one layer both sides may import, and already holds the feature-flag parser
and the availability type.

The selectors stay in ui/selectors/money-account-feature-flags.ts and now
import the parser; the flag name is written once, in the shared module, so
the UI selector and any background reader cannot end up looking at
different flags.
